+++ CHANGES.md
@@ -1,5 +1,5 @@
 ## Changelog
-  * Feature: New data channel - `futures_index` - demonstrated in ByBit
+  * Feature: Add stop callback when exiting loop, add stop method placeholder for base callbacks
 
 ### 1.6.2
   * Bugfix: Fix aggregated callbacks (Renko and OHLCV) when used with exchanges that support order types