+++ CHANGES.md
@@ -1,10 +1,10 @@
 ## Changelog
-  * Feature: Add stop callback when exiting loop, add stop method placeholder for base callbacks
 
 ### 1.6.2
   * Bugfix: Fix aggregated callbacks (Renko and OHLCV) when used with exchanges that support order types
   * Bugfix: Fix broken example/demo code
   * Feature: New data channel - `futures_index` - demonstrated in ByBit
+  * Feature: Add stop callback when exiting loop, add stop method placeholder for base callbacks
 
 ### 1.6.1 (2020-11-12)
   * Feature: New kwarg for exchange feed - `snapshot_interval` - used to control number of snapshot updates sent to client