+++ CHANGES.md
@@ -5,6 +5,7 @@
   * Bugfix: Fix broken example/demo code
   * Feature: New data channel - `futures_index` - demonstrated in ByBit
   * Feature: Add stop callback when exiting loop, add stop method placeholder for base callbacks
+  * Bugfix: Fix NBBO callback
 
 ### 1.6.1 (2020-11-12)
   * Feature: New kwarg for exchange feed - `snapshot_interval` - used to control number of snapshot updates sent to client