+++ CHANGES.md
@@ -6,6 +6,7 @@
   * Feature: New data channel - `futures_index` - demonstrated in ByBit
   * Feature: Add stop callback when exiting loop, add stop method placeholder for base callbacks
   * Bugfix: Fix NBBO callback
+  * Feature: Orderbook sequence number validation for HitBTC
 
 ### 1.6.1 (2020-11-12)
   * Feature: New kwarg for exchange feed - `snapshot_interval` - used to control number of snapshot updates sent to client