+++ CHANGES.md
@@ -7,6 +7,7 @@
   * Feature: Add stop callback when exiting loop, add stop method placeholder for base callbacks
   * Bugfix: Fix NBBO callback
   * Feature: Orderbook sequence number validation for HitBTC
+  * Feature: Kraken orderbook checksum support in Kraken
 
 ### 1.6.1 (2020-11-12)
   * Feature: New kwarg for exchange feed - `snapshot_interval` - used to control number of snapshot updates sent to client